requirements.txt
algoliasearch-django>=2.0,<3.0
aniso8601==7.0.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.3.0
blessings==1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.1
curtsies==0.3.10
cwcwidth==0.1.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-graphql-jwt==0.3.4
django-heroku==0.3.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-rest-swagger==2.2.0
djangorestframework==3.13.1
executing==0.8.3
fastdiff==0.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httpie==3.1.0
idna==3.3
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
openapi-codec==1.3.2
outcome==1.1.0
parso==0.8.3
pathspec==0.9.0
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
promise==2.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-slugify==6.1.2
pytz==2021.3
pyxdg==0.27
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
Rx==1.6.1
selenium==4.1.2
simplejson==3.17.6
singledispatch==3.7.0
six==1.16.0
snapshottest==0.6.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
traitlets==5.2.1.post0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.8
wasmer==1.1.0
wasmer_compiler_cranelift==1.1.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
wsproto==1.1.0
